[Update] cls_carro.py

+++ DEVinCar/classes/cls_carro.py
@@ -38,7 +38,7 @@ class Carro(Veiculo):
 
         placa = str(input("Digite a placa do carro: "))
         cor_nova = str(input(f"Digite a nova cor: "))
-        valor_novo = int(input(f"Digite o valor novo R$: "))
+        valor_novo = float(input(f"Digite o valor novo R$: "))
         
         for x in lista_carros:
